Added hand program in report - Improving SEA-622

+++ src/main/java/org/uario/seaworkengine/statistics/ShipTotal.java
@@ -13,7 +13,7 @@ public class ShipTotal implements Serializable {
 	private Double				containerRZ_TW_MCT;
 	private Double				containerRZ_TW_SWS;
 	private Double				handswork;
-	private Integer				handswork_program;
+	private Double				handswork_program;
 
 	private Integer				invoice_month;
 	private Double				menwork;
@@ -48,7 +48,7 @@ public class ShipTotal implements Serializable {
 		return this.handswork;
 	}
 
-	public Integer getHandswork_program() {
+	public Double getHandswork_program() {
 		return this.handswork_program;
 	}
 
@@ -116,7 +116,7 @@ public class ShipTotal implements Serializable {
 		this.handswork = handswork;
 	}
 
-	public void setHandswork_program(final Integer handswork_program) {
+	public void setHandswork_program(final Double handswork_program) {
 		this.handswork_program = handswork_program;
 	}
